==> notebooks/01_synthetic_data_creation.ipynb <==
import os, sys, pandas as pd, numpy as np, string, json, math
from pandas.io.json import json_normalize
from random import randrange, shuffle

DATA_FOLDER = "../data/"

SYN_REPORTS_FOLDER = "../data/outputs/synthetic_reports/"

def create_metric_names(num_metrics = 100, prefixes = ["volume_sales", "value_sales", "arpu", "user_name", 
                                                       "phone_num", "usage_tier", "total_calls", "total_sms", 
                                                       "total_data", "monthly_plan", "subscriptions", 
                                                       "roaming_calls", "roaming_data", "roaming_sms", 
                                                       "volume sales", "value sales", "arpu", "user name", 
                                                       "phone num", "usage tier", "total calls", "total sms", 
                                                       "total data", "monthly plan", "subscriptions", 
                                                       "roaming calls", "roaming data", "roaming sms",
                                                      "sales(volume)", "value sales", "arpu", "user name", 
                                                       "phone num", "usage tier", "total calls", "total sms", 
                                                       "total data", "monthly plan", "subscriptions", 
                                                       "roaming calls", "data(roaming)", "roaming sms"]):
    
    """
    Creates metric names based on the prefixes provided.
    create num_metrics number of metrics totally
    """
    
    pre_len = len(prefixes)
    
    num_replica = (num_metrics//pre_len) + 1
    
    metrics_list = []
    
    for prefix in prefixes:
        
        for rep_num in range(num_replica):
            
            if len(metrics_list) >= num_metrics:
                
                break
            
            metrics_list.append(prefix + "_" + str(rep_num))
            
    
    return metrics_list
    
    

with open('../data/raw/telstra.json') as jsonfile:
    
    telstra = json.load(jsonfile)
    
telstra_df = json_normalize(telstra)

telstra_df.head()

def create_metric_desc(df, num_metrics = 100):
    
    """
    Takes a sample data frame that should have "title" as column and 
    values in these columns should be text
    
    Returns num_metrics no of list of texts 
    
    """
    
    text_lists = df["title"].tolist()
    
    total_texts = len(text_lists)
    
    repeats = math.ceil(num_metrics / total_texts )
    
    text_lists = text_lists * repeats
    
    
    
    return text_lists[:num_metrics]

NUM_METRICS = 500

metric_names = create_metric_names(NUM_METRICS)

print(len(metric_names))

metric_descriptions = create_metric_desc(telstra_df, num_metrics=NUM_METRICS)

print(len(metric_descriptions))

metrics_catalog_df = pd.DataFrame({"metrics": metric_names, "description": metric_descriptions})

metrics_catalog_df.sample(5)

metrics_catalog_df.to_csv("{}metrics_catalog_synthetic_v1_19feb20.csv".format(DATA_FOLDER + "outputs/"), index=False)

#### Reports Creation

column_names = create_metric_names(NUM_METRICS + 500, prefixes=["nsw", "qld", "sydney", "melbourne", 
                                                       "brisbane", "perth", "victoria", "canberra", 
                                                       "adelaid", "gold coast", "auckland"])

pd.options.display.max_rows = 1000

def _delete_files_from_folder(foldername = SYN_REPORTS_FOLDER):
    
    abs_path = os.path.abspath(SYN_REPORTS_FOLDER)
    
    print("Deleting files from {}".format(abs_path))
    
    filenames = os.listdir(abs_path)
    
    counter = 0
    
    for filename in filenames:
        
        abs_filename = os.path.join(abs_path, filename)

        try:
            os.remove(abs_filename)
            
            counter += 1
            
        except Exception as e:
            
            print("unable to delete "+ filename + "_" +str(e))
            
            pass
        
    print("{}/{} files deleted.".format(str(counter), str(len(filenames))))
    
    return

def create_reports(column_names, num_reports = 10, num_rows = 10):
    
    """
    Delete currently present reports, then create num_reports # of  excel files with num_rows
    
    Column names for eacg file will be randomly sampled from provided column names
    
    num cols per file is randomly varied to be around (len(column_names)//num_reports + 5)
    
    similarly number of rows are randomly varied for each report
    
    A string column is randomly added to each report. Values in this column will be randomly the column names themselves
    """
    
    num_columns = len(column_names)
    
    num_cols_rep = math.ceil(num_columns // num_reports) + 5
    
    reports_list = []
    
    """
    First delete all files in synthetic reports folder
    """
    _delete_files_from_folder(SYN_REPORTS_FOLDER)
    
    for rep in range(num_reports):
        
        num_columns_curr = randrange(2, num_cols_rep)
        
        string_col_num = randrange(1, num_columns_curr)
        
        num_rows_curr = randrange(2, num_rows + 5)
        
        col_nums = np.random.randint(1, num_columns, num_columns_curr).tolist()
        
        col_names = [column_names[x]for x in col_nums]
        
        string_colname = col_names[string_col_num]
        
        curr_report_df = pd.DataFrame(columns=col_names)
        
        for col in col_names:
            
            if col == string_colname:
                
                idxs = np.random.randint(1, num_columns, num_rows_curr).tolist()
                
                values = [column_names[x] for x in idxs]
                
                #print(col)
                
            else:
                
                values = np.random.rand(num_rows_curr).tolist()
            
            curr_report_df[col] = values
            
        print("{} report(s) created".format(str(rep + 1)))
        
        reports_list.append(curr_report_df)

        curr_report_df.to_excel("{}reports_synthetic_{}_v1_19feb20.xlsx".format(SYN_REPORTS_FOLDER, str(rep)), 
                                      index=False)
    
    return reports_list

df_list = create_reports(column_names, num_reports = 204, num_rows = 100)

def create_cols_metric_mapping(colnames, metric_names):
    
    """
    Creates a synthetic mapping b/w colnames and metric names
    
    Assumes, more than 1 column could be mapped to a metric
    """
    
    # Introducing duplicates in metrics
    metric_names_duplicated = metric_names * 2
    
    #Random shuffle
    shuffle(metric_names_duplicated)
    
    #Sample from this duplicated randomly shuffled metric names and assign them against columns
    
    num_cols = len(colnames)
    
    sampled_metric_names = metric_names_duplicated[:num_cols]
    
    col_met_mapping = pd.DataFrame({"column_names":colnames, "metric_names":sampled_metric_names})   
    
    
    col_met_mapping.to_csv("{}col_met_mapping_v1_19feb20.csv".format(DATA_FOLDER+"outputs/"), index=False)
    
    return col_met_mapping

col_met_mapping = create_cols_metric_mapping(colnames=column_names, metric_names=metric_names)

col_met_mapping.head(5)

col_met_mapping.metric_names.duplicated().sum()

#### Users - Files -Owners mapping

names_list = telstra_df["user"].unique().tolist()

print(len(names_list))

names_list

report_names = os.listdir(SYN_REPORTS_FOLDER)

def create_user_rep_owner_mapping(names_list, report_names):
    
    """
    Creates report - owners mapping
    and report - users mapping
    
    A owner can be mapped to multiple reports
    
    Each report will have multiple users in a many to many fashion
    """
    
    num_reports = len(report_names)
    
    num_names = len(names_list)
    
    #Duplicate usernames and randomly assign them to reports
    
    num_names_rep = math.ceil(num_reports / num_names) + 50
    
    names_list_dup = names_list * num_names_rep
    
    shuffle(names_list_dup)
    
    owner_names = names_list_dup[:num_reports]
    
    #Duplicate report names and randomly assign them to duplicated random users
    
    num_rep_rep = math.ceil(num_names / num_reports) + 3
    
    reports_list_dup = report_names * num_rep_rep
    
    shuffle(reports_list_dup)
    
    names_mult_factor = math.ceil(len(reports_list_dup) // num_names) + 1
    
    names_list_dup2 = names_list * names_mult_factor
    
    shuffle(names_list_dup2)
    
    user_names_list = names_list_dup2[: len(reports_list_dup)]
    
    owners_df = pd.DataFrame({"Reports": report_names, "Owners":owner_names})
    
    users_df = pd.DataFrame({"Reports": reports_list_dup, "Users":user_names_list})
    
    users_df.drop_duplicates(inplace=True)
    
    users_df.sort_values("Reports", inplace=True)
    
    owners_df.sort_values("Reports", inplace=True)
    
    users_df.reset_index(drop=True, inplace=True)
    
    print(users_df.shape)
    
    owners_df.to_csv("{}rep_owner_mapping_v1_19feb20.csv".format(DATA_FOLDER + "outputs/"), index=False)
    
    users_df.to_csv("{}rep_users_mapping_v1_19feb20.csv".format(DATA_FOLDER + "outputs/"), index=False)
    
    
    return [users_df, owners_df]

users_df, owners_df = create_user_rep_owner_mapping(names_list, report_names)

users_df

print(owners_df.Owners.duplicated().sum())

owners_df
