==> notebooks/03_query_api_experiments.ipynb <==
import os, sys, pandas as pd, numpy as np, string, json, math
from random import randrange, shuffle
from elasticsearch import Elasticsearch

es = Elasticsearch()

def search_reports(kw="sales", 
                   es_connector=Elasticsearch(),
                   index="test",
                   field="metric_names"
                  
                  ):
    
    """
    Searches for report names using metric name keyword
    """
    
    
    
    fields_list = []
    
    fields_list.append(field)
    
    #search_term_proc = "*" + kw + "*" #adding wildcard
    
    search_term_proc = kw #adding wildcard
    
    
    
    results = es_connector.search(index=index, 
          body={
              "size":10000,
              "query": 
                {"query_string" : { 
                    "query" : search_term_proc,
                    "fields" : ["metric_names*"],
                    
                }
                }
               })
    return pd.json_normalize(results["hits"]["hits"])

search_reports(kw="volume sales_4", field="metric_names.raw", index="test2")#.sort_values(["_score", "_source.metric_names"], ascending=False)

def search_reports(kw="sales", 
                   es_connector=Elasticsearch(),
                   index="test",
                   field="metric_names"
                  
                  ):
    
    """
    Searches for report names using metric name keyword
    """
    
    
    
    fields_list = []
    
    fields_list.append(field)
    
    kw = kw.strip()
    
    search_term_proc = "*" + kw + "*" #adding wildcard
    
    #search_term_proc = kw #adding wildcard
    
    st2 = "*"+ (" ".join(kw.split("_"))).strip() +"*"
    
    custom_fname = field + ".custom"
    
    std_fname = field + ".standard"
    
    raw_fname = field + ".raw"
    
        
    results = es_connector.search(index=index, 
                                  body={ 
                                   "size":10000,
                                   "query":{ 
                                      "bool":{ 
                                         "should":[ 
                                            { 
                                               "query_string":{ 
                                                  "query":search_term_proc,
                                                  "fields":[ 
                                                     custom_fname
                                                  ],
                                                  "boost":1
                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":kw,
                                                  "fields":[ 
                                                     custom_fname
                                                  ],
                                                  "boost":1,
                                                  "fuzziness":100,
                                                  "fuzzy_max_expansions":1000
                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":search_term_proc,
                                                  "fields":[ 
                                                     raw_fname
                                                  ],
                                                  "boost":3
                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":search_term_proc,
                                                  "fields":[ 
                                                     std_fname
                                                  ],
                                                  "boost":1.5,
                                                  "fuzziness":10,

                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":st2,
                                                  "fields":[ 
                                                     custom_fname
                                                  ],
                                                  "boost":1
                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":st2,
                                                  "fields":[ 
                                                     raw_fname
                                                  ],
                                                  "boost":3
                                               }
                                            },
                                            { 
                                               "query_string":{ 
                                                  "query":st2,
                                                  "fields":[ 
                                                     std_fname
                                                  ],
                                                  "boost":1.5,
                                                  "fuzziness":10,

                                               }
                                            }
                                         ]
                                      }
                                   }
                                })
    return pd.json_normalize(results["hits"]["hits"])

search_reports(kw="total calls", field="metric_names", index="test2").sort_values(["_score", "_source.metric_names"], ascending=False)

search_reports(kw="total calls", field="metric_names", index="test2").sort_values(["_score", "_source.metric_names"], ascending=False)["_source.metric_names"].nunique()

df = search_reports(kw="sales", field="metric_names", index="test2")

df["_source.metric_names"].nunique()

pd.options.display.max_rows = 1000

es = Elasticsearch()

es.indices.analyze(index="test2", body={
  "analyzer" : "metric_names_analyzer",
  "text" : "arp_"
})

es.indices.get_settings("test2")

es.indices.get_mapping("test2")

es.indices.open("test")

es.indices.put_settings(
    index="test",
    body={
  "settings": {
    "analysis": {
      "analyzer": {
        "metric_names_analyzer": {
          "tokenizer": "underscore_tokenizer"
        }
      },
      "tokenizer": {
        "underscore_tokenizer": {
          "type": "pattern",
          "pattern": "_"
        }
      }
    }
  }
}
)

{'mappings': {
    'properties': {
        'colname': {'type': 'text',
                    'fields': {'metric_names': {'type': 'text', "analyzer":"metric_names_analyzer"}}},
        'metric_names': {'type': 'text',
                         'fields': {'metric_names': {'type': 'text', "analyzer":"metric_names_analyzer"}}},
        'repname': {'type': 'text',
                    'fields': {'metric_names': {'type': 'text', "analyzer":"metric_names_analyzer"}}}
    }
}
}

es.indices.put_mapping(
    index="test",
    doc_type="my-type",
    body={
        "properties": {  
            "date": {"type":"date"},
            "word_data": {"type": "text"},
            "word_count": {"type": "integer"}
        }
    }
)

es.get(index="test2", id=1)

es.search(index="test", body={"query": 
{"fuzzy" : { "metric_names" : 
{"value": "roaming_calls",
"fuzziness": 20
}}}})

"analysis": {
      "analyzer": {
        "my_analyzer": {
          "tokenizer": "my_tokenizer"
        }
      },
      "tokenizer": {
        "my_tokenizer": {
          "type": "simple_pattern_split",
          "pattern": "_"
        }
      }
    }

es.search(index="test", 
          body={"query": 
                {"query_string" : { 
                    "query" : "*sales*",
                    "fields" : ["metric_names"]
                           }
                }
               })

,"fields":{
                              "raw": { 
                                "type":     "keyword",
                                "analyzer": "standard",
                                "search_analyzer":"standard",
                              }
                   }

index_map_setts = {'mappings': {
    'properties': {
        'colname': {'type': 'text',
                    "fields":{
                              "raw": { 
                                "type":     "keyword"
                              },
                           "standard": { 
                                "type":     "text",
                               "analyzer":"standard",
                               "search_analyzer":"standard"
                              },
                           "custom": { 
                                "type":     "text",
                               "analyzer":"metric_names_analyzer",
                               "search_analyzer":"metric_names_analyzer"
                              }
                   }},
        'metric_names': {'type': 'text',
                         "fields":{
                              "raw": { 
                                "type":     "keyword"
                              },
                           "standard": { 
                                "type":     "text",
                               "analyzer":"standard",
                               "search_analyzer":"standard"
                              },
                            "custom": { 
                                "type":     "text",
                               "analyzer":"metric_names_analyzer",
                               "search_analyzer":"metric_names_analyzer"
                              }
                   }},
        'repname': {'type': 'text',
                    "fields":{
                              "raw": { 
                                "type":     "keyword"
                              },
                           "standard": { 
                                "type":     "text",
                               "analyzer":"standard",
                               "search_analyzer":"standard"
                              },
                           "custom": { 
                                "type":     "text",
                               "analyzer":"metric_names_analyzer",
                               "search_analyzer":"metric_names_analyzer"
                              }
                   }},
    }
},
"settings": {
    "analysis": {
      "analyzer": {
        "metric_names_analyzer": {
          "tokenizer": "underscore_tokenizer"
        }
      },
      "tokenizer": {
        "underscore_tokenizer": {
          "type": "pattern",
          "pattern": "_"
        }
      }
    }
  }
 
}


if not es.indices.exists(index="test2"):
    
    print("done")
    
    es.indices.create(index="test2",
         body=index_map_setts)

es.indices.delete("test2")

es.indices.create(index="test2",
         body=index_map_setts)

es.indices.get_mapping("test2")
